+++ backend/src/baserow/contrib/database/api/views/errors.py
@@ -1,4 +1,8 @@
-from rest_framework.status import HTTP_400_BAD_REQUEST, HTTP_404_NOT_FOUND
+from rest_framework.status import (
+    HTTP_400_BAD_REQUEST,
+    HTTP_401_UNAUTHORIZED,
+    HTTP_404_NOT_FOUND,
+)
 
 
 ERROR_VIEW_DOES_NOT_EXIST = (
@@ -80,4 +84,9 @@ ERROR_CANNOT_SHARE_VIEW_TYPE = (
     "ERROR_CANNOT_SHARE_VIEW_TYPE",
     HTTP_400_BAD_REQUEST,
     "This view type does not support sharing.",
+)
+ERROR_NO_AUTHORIZATION_TO_PUBLICLY_SHARED_VIEW = (
+    "ERROR_NO_AUTHORIZATION_TO_PUBLICLY_SHARED_VIEW",
+    HTTP_401_UNAUTHORIZED,
+    "The user does not have the permissions to see this password protected shared view.",
 )